delete RaoPtdfParameters
