+++ ra-optimisation/rao-api/src/test/java/com/farao_community/farao/rao_api/RaoParametersTest.java
@@ -6,6 +6,7 @@
  */
 package com.farao_community.farao.rao_api;
 
+import com.farao_community.farao.commons.FaraoException;
 import com.google.auto.service.AutoService;
 import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
@@ -13,11 +14,17 @@ import com.powsybl.commons.config.InMemoryPlatformConfig;
 import com.powsybl.commons.config.MapModuleConfig;
 import com.powsybl.commons.config.PlatformConfig;
 import com.powsybl.commons.extensions.AbstractExtension;
+import com.powsybl.iidm.network.Country;
+import org.apache.commons.lang3.tuple.ImmutablePair;
+import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
 import java.nio.file.FileSystem;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
 import static org.junit.Assert.*;
@@ -100,6 +107,21 @@ public class RaoParametersTest {
     }
 
     @Test
+    public void checkPtdfConfig() {
+        MapModuleConfig moduleConfig = platformCfg.createModuleConfig("rao-parameters");
+        moduleConfig.setStringListProperty("ptdf-boundaries", new ArrayList<>(Arrays.asList("FR-ES", "ES-PT")));
+        moduleConfig.setStringProperty("ptdf-sum-lower-bound", Objects.toString(5.0));
+
+        RaoParameters parameters = new RaoParameters();
+        RaoParameters.load(parameters, platformCfg);
+
+        assertEquals(5, parameters.getPtdfSumLowerBound(), 1e-6);
+        assertEquals(2, parameters.getPtdfBoundaries().size());
+        assertTrue(parameters.getPtdfBoundaries().contains(new ImmutablePair<>(Country.FR, Country.ES)));
+        assertTrue(parameters.getPtdfBoundaries().contains(new ImmutablePair<>(Country.ES, Country.PT)));
+    }
+
+    @Test
     public void checkRelativeMarginConfig() {
         MapModuleConfig moduleConfig = platformCfg.createModuleConfig("rao-parameters");
         moduleConfig.setStringProperty("negative-margin-objective-coefficient", Objects.toString(100.0));
@@ -169,5 +191,48 @@ public class RaoParametersTest {
         public Class<? super DummyExtension> getExtensionClass() {
             return DummyExtension.class;
         }
+    }
+
+    @Test
+    public void testSetBoundariesFromCountryCodes() {
+        RaoParameters parameters = new RaoParameters();
+        List<String> stringBoundaries = new ArrayList<>(Arrays.asList("FR-ES", "ES-PT"));
+        parameters.setPtdfBoundariesFromCountryCodes(stringBoundaries);
+        assertEquals(2, parameters.getPtdfBoundaries().size());
+        assertTrue(parameters.getPtdfBoundaries().contains(new ImmutablePair<>(Country.FR, Country.ES)));
+        assertTrue(parameters.getPtdfBoundaries().contains(new ImmutablePair<>(Country.ES, Country.PT)));
+    }
+
+    private void testWrongBoundary(String boundary) {
+        RaoParameters parameters = new RaoParameters();
+        List<String> stringBoundaries = new ArrayList<>(Arrays.asList(boundary));
+        parameters.setPtdfBoundariesFromCountryCodes(stringBoundaries);
+    }
+
+    @Test (expected = FaraoException.class)
+    public void testSetBoundariesFromCountryCodesException1() {
+        testWrongBoundary("FRANCE-SPAIN");
+    }
+
+    @Test (expected = FaraoException.class)
+    public void testSetBoundariesFromCountryCodesException2() {
+        testWrongBoundary("FR-ES-");
+    }
+
+    @Test (expected = FaraoException.class)
+    public void testSetBoundariesFromCountryCodesException3() {
+        testWrongBoundary("FR-YY");
+    }
+
+    @Test
+    public void testGetBoundariesFromString() {
+        RaoParameters parameters = new RaoParameters();
+        List<Pair<Country, Country>> countryBoundaries = new ArrayList<>(
+                Arrays.asList(new ImmutablePair<>(Country.BE, Country.FR), new ImmutablePair<>(Country.DE, Country.AT))
+        );
+        parameters.setPtdfBoundaries(countryBoundaries);
+        assertEquals(2, parameters.getPtdfBoundariesAsString().size());
+        assertTrue(parameters.getPtdfBoundariesAsString().contains("BE-FR"));
+        assertTrue(parameters.getPtdfBoundariesAsString().contains("DE-AT"));
     }
 }
