+++ ra-optimisation/rao-api/src/test/resources/RaoParameters.json
@@ -14,6 +14,8 @@
   "mnec-violation-cost" : 10.0,
   "mnec-constraint-adjustment-coefficient" : 0.0,
   "negative-margin-objective-coefficient" : 1000.0,
+  "ptdf-boundaries" : [ ],
+  "ptdf-sum-lower-bound" : 0.01,
   "sensitivity-parameters" : {
     "version" : "1.0",
     "load-flow-parameters" : {
