+++ ra-optimisation/rao-api/src/test/resources/RaoParametersSet.json
@@ -14,6 +14,8 @@
   "mnec-violation-cost" : 20.0,
   "mnec-constraint-adjustment-coefficient" : 3.0,
   "negative-margin-objective-coefficient" : 100.0,
+  "ptdf-boundaries" : [ "FR-ES", "ES-PT" ],
+  "ptdf-sum-lower-bound" : 0.05,
   "sensitivity-parameters" : {
     "version" : "1.0",
     "load-flow-parameters" : {
