+++ ra-optimisation/rao-commons/src/test/java/com/farao_community/farao/rao_commons/objective_function_evaluator/MinMarginObjectiveFunctionTest.java
@@ -17,7 +17,6 @@ import com.farao_community.farao.data.crac_result_extensions.ResultVariantManage
 import com.farao_community.farao.rao_api.RaoParameters;
 import com.farao_community.farao.rao_commons.RaoData;
 import com.farao_community.farao.rao_commons.RaoInputHelper;
-import com.farao_community.farao.rao_commons.RaoPtdfParameters;
 import com.farao_community.farao.sensitivity_analysis.SystematicSensitivityResult;
 import com.powsybl.iidm.network.Network;
 import org.junit.Test;
@@ -58,8 +57,7 @@ public class MinMarginObjectiveFunctionTest {
         raoParameters.setMnecAcceptableMarginDiminution(mnecAcceptableMarginDiminution);
         raoParameters.setMnecViolationCost(mnecViolationCost);
         raoParameters.setObjectiveFunction(objectiveFunction);
-        raoParameters.addExtension(RaoPtdfParameters.class, new RaoPtdfParameters());
-        raoParameters.getExtension(RaoPtdfParameters.class).setPtdfSumLowerBound(ptdfSumLowerBound);
+        raoParameters.setPtdfSumLowerBound(ptdfSumLowerBound);
 
         minRelativeMarginObjectiveFunction = new MinMarginObjectiveFunction(raoParameters);
         crac.newCnec().setId("MNEC1 - initial-instant - preventive")
