--- ra-optimisation/rao-commons/src/test/resources/RaoParametersWithBoundaries.json
@@ -1,35 +1,0 @@
-{
-  "version" : "1.0",
-  "objective-function" : "MAX_MIN_MARGIN_IN_MEGAWATT",
-  "max-number-of-iterations" : 10,
-  "pst-penalty-cost" : 0.01,
-  "pst-sensitivity-threshold" : 0.0,
-  "sensitivity-fallback-over-cost" : 0.0,
-  "rao-with-loop-flow-limitation" : false,
-  "loop-flow-approximation" : "FIXED_PTDF",
-  "loop-flow-constraint-adjustment-coefficient" : 0.0,
-  "loop-flow-violation-cost" : 0.0,
-  "sensitivity-parameters" : {
-    "version" : "1.0",
-    "load-flow-parameters" : {
-      "version" : "1.3",
-      "voltageInitMode" : "UNIFORM_VALUES",
-      "transformerVoltageControlOn" : false,
-      "phaseShifterRegulationOn" : false,
-      "noGeneratorReactiveLimits" : false,
-      "twtSplitShuntAdmittance" : false,
-      "simulShunt" : false,
-      "readSlackBus" : false,
-      "writeSlackBus" : false
-    }
-  },
-  "extensions" : {
-    "RaoPtdfParameters" : {
-      "boundaries" : [
-        "FR-ES",
-        "ES-PT"
-      ],
-      "ptdf-sum-lower-bound" : 0.07
-    }
-  }
-}
