--- ra-optimisation/rao-commons/src/test/resources/RaoPtdfParameters.json
@@ -1,40 +1,0 @@
-{
-  "version" : "1.0",
-  "objective-function" : "MAX_MIN_MARGIN_IN_MEGAWATT",
-  "max-number-of-iterations" : 10,
-  "pst-penalty-cost" : 0.01,
-  "pst-sensitivity-threshold" : 0.0,
-  "sensitivity-fallback-over-cost" : 0.0,
-  "rao-with-loop-flow-limitation" : false,
-  "loop-flow-approximation" : "FIXED_PTDF",
-  "loop-flow-constraint-adjustment-coefficient" : 0.0,
-  "loop-flow-violation-cost" : 0.0,
-  "loop-flow-countries" : [ ],
-  "mnec-acceptable-margin-diminution" : 50.0,
-  "mnec-violation-cost" : 10.0,
-  "mnec-constraint-adjustment-coefficient" : 0.0,
-  "negative-margin-objective-coefficient" : 1000.0,
-  "sensitivity-parameters" : {
-    "version" : "1.0",
-    "load-flow-parameters" : {
-      "version" : "1.4",
-      "voltageInitMode" : "UNIFORM_VALUES",
-      "transformerVoltageControlOn" : false,
-      "phaseShifterRegulationOn" : false,
-      "noGeneratorReactiveLimits" : false,
-      "twtSplitShuntAdmittance" : false,
-      "simulShunt" : false,
-      "readSlackBus" : false,
-      "writeSlackBus" : false,
-      "dc" : false,
-      "distributedSlack" : true,
-      "balanceType" : "PROPORTIONAL_TO_GENERATION_P_MAX"
-    }
-  },
-  "extensions" : {
-    "RaoPtdfParameters" : {
-      "boundaries" : [ "FR-ES", "ES-PT" ],
-      "ptdf-sum-lower-bound" : 0.05
-    }
-  }
-}
